+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/list/ListDownloadServer.kt
@@ -1,0 +1,10 @@
+package au.com.shiftyjelly.pocketcasts.servers.list
+
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface ListDownloadServer {
+
+    @GET("{listId}.json")
+    suspend fun getPodcastList(@Path("listId") listId: String): PodcastList
+}